+++ inst/examples/ex-jackknife.R
@@ -1,0 +1,7 @@
+x <- rnorm(20)
+
+## Jackknife
+jackknife(x, do = mean) # Sample mean
+
+## Get the leave-one-out values instead of summary
+jackknife(x, do = mean, f = function(x) { x })