+++ inst/examples/ex-resample.R
@@ -1,17 +1,7 @@
+## Uniform distribution
 x <- rnorm(20)
+resample_uniform(x, n = 10)
 
-## Bootstrap
-bootstrap(x, do = mean, n = 100)
-
-## Estimate the 25th and 95th percentiles
-quant <- function(x) { quantile(x, probs = c(0.25, 0.75)) }
-bootstrap(x, n = 100, do = mean, f = quant)
-
-## Get the n bootstrap values
-bootstrap(x, n = 100, do = mean, f = function(x) { x })
-
-## Jackknife
-jackknife(x, do = mean) # Sample mean
-
-## Get the leave-one-out values instead of summary
-jackknife(x, do = mean, f = function(x) { x })
+## Multinomial distribution
+x <- sample(1:100, 20, TRUE)
+resample_multinomial(x, n = 10)